api/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
ultralytics[cpu]
python-multipart

